compile_test wasn't testing what I wanted: because it ran in the main ctemplate source tree, it had access to files that weren't going to be installed, when the whole point it to test this can compile against the installed ctemplate.
To resolve this, I added a new rule that's executed at install
time (or the 'install' done during 'make distcheck').  I've
verified that this finds the problems with the current
ctemplate distribution that compile_test was supposed to
capture, and that are reported in
http://code.google.com/p/google-ctemplate/issues/detail?id=79

While in the area, add a few files that aren't in the dist
tarball but should be.  MOE noticed these.

R=jad
DELTA=11  (10 added, 0 deleted, 1 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2766


git-svn-id: http://ctemplate.googlecode.com/svn/trunk@73 cb39b542-eb2e-2a05-867b-f1791dd5f601