A user reported a bug that configure should die when g++ isn't installed, but wasn't: http://code.google.com/p/google-perftools/issues/detail?id=217
This turned out to be a bug in autoconf:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=357378

The workaround is simple: check for CXX before checking for
CC.  This means that if g++ is installed but gcc isn't, we
won't die (at cc-checking time), but I tested, and the
configure script dies later.  In any case, it seems unlikely
someone would have a c++ compiler installed but not a c
compiler.

This fixes the 4 opensource projects I work on that are
susceptible to this.

R=iant
DELTA=1437  (694 added, 694 deleted, 49 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3221


git-svn-id: http://ctemplate.googlecode.com/svn/trunk@82 cb39b542-eb2e-2a05-867b-f1791dd5f601